+++ lib/hcloud/resources/floating_ip.rb
@@ -22,6 +22,9 @@ module HCloud
   #     HCloud::FloatingIP.where(name: "my_floating_ip")
   #     # => #<HCloud::FloatingIP id: 1, ...>
   #
+  #     HCloud::FloatingIP.where(label_selector: { environment: "production" })
+  #     # => #<HCloud::FloatingIP id: 1, ...>
+  #
   # == Find floating IP by ID
   #
   #     HCloud::FloatingIP.find(1)