+++ lib/hcloud/resources/placement_group.rb
@@ -22,6 +22,9 @@ module HCloud
   #     HCloud::PlacementGroup.where(name: "my_placement_group")
   #     # => #<HCloud::PlacementGroup id: 2, ...>
   #
+  #     HCloud::PlacementGroup.where(label_selector: { environment: "production" })
+  #     # => #<HCloud::PlacementGroup id: 2, ...>
+  #
   # == Find placement group by ID
   #
   #     HCloud::PlacementGroup.find(2)