+++ lib/hcloud/resources/volume.rb
@@ -25,6 +25,9 @@ module HCloud
   #     HCloud::Volume.where(status: "available")
   #     # => #<HCloud::Volume id: 1, ...>
   #
+  #     HCloud::Volume.where(label_selector: { environment: "production" })
+  #     # => #<HCloud::Volume id: 1, ...>
+  #
   # == Find volume by ID
   #
   #     HCloud::Volume.find(1)