+++ spec/hcloud/concerns/queryable_spec.rb
@@ -67,6 +67,16 @@ RSpec.describe HCloud::Queryable do
 
       expect(models.count).to eq 1
     end
+
+    it "returns resources selected by label" do
+      stub_request(:get, "https://api.hetzner.cloud/v1/examples")
+        .with(query: { page: 1, per_page: 50, label_selector: "foo=bar,baz=bat" })
+        .to_return(body: { examples: [resource.attributes.merge(id: 1, labels: { foo: "bar", baz: "bat" })], meta: { pagination: { total_entries: 1 } } }.to_json)
+
+      models = ExampleResource.all.where(label_selector: { foo: "bar", baz: "bat" })
+
+      expect(models.count).to eq 1
+    end
   end
 
   describe ".first, .last, .count, .where, .sort, .each, .empty?, .any?" do