+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `Action#resources` now returns a list of resources instead of a list of hashes
 - Separate server protection entity from other protection entity (only server protection includes `rebuild`)
 - Add shorthand methods for `.first`, `.last`, `.count`, `.where`, `.sort`, `.each`, and `.empty?` on resource
+- Add `label_selector:` argument to `HCloud::Collection`
 
 ## HCloud v1.4.0 (2023-01-22)